feat(game): tile-actions modal opens directly from world map
Clicking a tile on /game/tiles now opens an inline modal with the same
actions as the dedicated tile page (assign land type, build units, arm
defense spell, attack), instead of navigating away. Action responses
flow back into the localStorage map cache so the map repaints with
the result without a refresh.

- Extract OwnTilePanel / EnemyTilePanel / Stat / UNIT_TYPES into
  app/game/tiles/_components/tile-action-panels.tsx, widening the tile
  prop type from GameTile to MapTile (only fields available on both
  are referenced).
- New TileActionsModal: backdrop dismiss, Esc to close, links to
  /game/upgrades and the full tile page for cases the modal doesn't
  cover.
- Click handler on the world map opens the modal for any owned tile
  (own or enemy). Unrevealed / unowned tiles remain hover-only.
- Hover hint now reads "click to manage" / "click to attack"
  appropriately.
